deployment/requirements.txt
# Adelino deployment dependencies
numpy>=1.24.0
websocket-client>=1.6.0
onnxruntime>=1.16.0
# torch is needed only if using TorchScript (.pt) policies instead of ONNX
# torch>=2.0.0
